Added TTD org issuer support for PlatformAccessTokens
- [#827](https://github.com/Altinn/altinn-access-management/issues/827)

Test serviceowner Testdepartementet (ttd) has been added as a valid issuer of PlatformAccessTokens, in order to support both manual and automated testing of Altinn 3 Platform APIs provided to App-backend solutions running in a service owners own cluster.

These tokens are use the serviceowner's org code as issuer (iss claim) and are signed with a private certificate available for all apps in the cluster.